+++ source/include/Ostap/Quantiles.h
@@ -12,6 +12,7 @@
 // =============================================================================
 #include "Ostap/Math.h"
 #include "Ostap/MakeArray.h"
+#include "Ostap/QuantileTypes.h"
 // =============================================================================
 namespace Ostap
 {
@@ -67,7 +68,7 @@ namespace Ostap
         { return (*this) ( first , last , k * 1.0 / N  ) ; } ;
         return Ostap::Math::make_array ( q , std::make_index_sequence<N+1>() )  ;
       } ;
-      // 1-quantiles: min & max
+      // 1-quantiles:   min, max
       template <class ITERATOR,
                 typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
                 typename std::enable_if<std::is_convertible<value_type,double>::value,bool>::type = true >
@@ -76,7 +77,7 @@ namespace Ostap
       ( ITERATOR first , 
         ITERATOR last  ) const
       { return this->template quantiles_<1> ( first  , last ) ;	 }
-      // 2-quantiles: min,median,max
+      // 2-quantiles: min,median & max
       template <class ITERATOR,
                 typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
                 typename std::enable_if<std::is_convertible<value_type,double>::value,bool>::type = true >
@@ -85,7 +86,7 @@ namespace Ostap
       ( ITERATOR first , 
         ITERATOR last  ) const
       { return this->template quantiles_<2> ( first  , last ) ;}
-      /// 3-quantiles:
+      /// 3-quantiles: min,q1,q2,max 
       template <class ITERATOR,
                 typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
                 typename std::enable_if<std::is_convertible<value_type,double>::value,bool>::type = true  >
@@ -94,7 +95,7 @@ namespace Ostap
       ( ITERATOR first , 
         ITERATOR last  ) const
       { return this->template quantiles_<3> ( first  , last ) ;}
-      /// 4-quantiles:
+      /// 4-quantiles: min,q1,q2,q3,max
       template <class ITERATOR,
                 typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
                 typename std::enable_if<std::is_convertible<value_type,double>::value,bool>::type = true  >
@@ -241,7 +242,7 @@ namespace Ostap
       ( ITERATOR first ,
         ITERATOR last  ) const 
       {
-        if  ( first == last ) 
+        if  ( first == last )
           { this->throw_exception ( "Input data cannot be empty!" , __FILE__ , __LINE__ ) ; }
         if ( this->m_check && !std::is_sorted ( first , last ) )
           { this->throw_exception ( "Input data must be sorted!"  , __FILE__ , __LINE__ ) ; }
@@ -268,32 +269,18 @@ namespace Ostap
      */
     class HyndmanFan : public QuantileMixin<HyndmanFan>
     {
-      //========================================================================
-    public:
       // =======================================================================
-      enum QuantileType
-        {
-          One = 1 ,
-          Two     ,
-          Three   ,
-          Four    ,
-          Five    ,
-          Six     ,
-          Seven   ,
-          Eight   ,
-          Nine    , 
-        } ;
-      // =======================================================================
-    public:
+    public:      
       // =======================================================================
       /// constructor
       HyndmanFan 
-      ( const QuantileType t     = Eight , 
-        const bool         check = false ) ; 
+      ( const Ostap::QuantileTypes::HyndmanFanType t =
+        Ostap::QuantileTypes::HyndmanFanType::Eight , 
+        const bool check = false ) ; 
       // =======================================================================
     public:
       // =======================================================================
-      /** get quanitile 
+      /** get quantiile 
        *  @param firts (INPUT) begin-iteratir for the input sequence
        *  @param last  (INPUT) end-iterator for the innut sequence 
        *  @param p     (INPUT) probability   0<p<1 
@@ -320,14 +307,14 @@ namespace Ostap
         if  ( 1 <= p ) { std::advance ( first , N - 1 ) ; return *first ; }
         //
         const double h = 
-          ( One   == m_t ) ?   N * p                  :
-          ( Two   == m_t ) ?   N * p + 0.5            :
-          ( Three == m_t ) ?   N * p - 0.5            : 
-          ( Four  == m_t ) ?   N * p                  :
-          ( Five  == m_t ) ?   N * p + 0.5            : 
-          ( Six   == m_t ) ?   N * p - 0.5            :
-          ( Seven == m_t ) ? ( N - 1    ) * p + 1     :
-          ( Eight == m_t ) ? ( N + 1./3 ) * p + 1./3  :
+          ( Ostap::QuantileTypes::HyndmanFanType::One   == m_t ) ?   N * p                  :
+          ( Ostap::QuantileTypes::HyndmanFanType::Two   == m_t ) ?   N * p + 0.5            :
+          ( Ostap::QuantileTypes::HyndmanFanType::Three == m_t ) ?   N * p - 0.5            : 
+          ( Ostap::QuantileTypes::HyndmanFanType::Four  == m_t ) ?   N * p                  :
+          ( Ostap::QuantileTypes::HyndmanFanType::Five  == m_t ) ?   N * p + 0.5            : 
+          ( Ostap::QuantileTypes::HyndmanFanType::Six   == m_t ) ?   N * p - 0.5            :
+          ( Ostap::QuantileTypes::HyndmanFanType::Seven == m_t ) ? ( N - 1    ) * p + 1     :
+          ( Ostap::QuantileTypes::HyndmanFanType::Eight == m_t ) ? ( N + 1./3 ) * p + 1./3  :
           ( N + 0.25 ) * p + 0.375 ;
         //
         /// clamped and adjusted for zero-indexed  
@@ -342,13 +329,13 @@ namespace Ostap
         //
         // first three cases explicitley 
         //
-        if ( One == m_t )
+        if ( Ostap::QuantileTypes::HyndmanFanType::One == m_t )
           {
             const std::size_t nn = Ostap::Math::round_up ( hh ) ;
             if ( nn ) { std::advance ( first , nn ) ; } 
             return *first ;
           }
-        else if ( Two == m_t )
+        else if ( Ostap::QuantileTypes::HyndmanFanType::Two == m_t )
           {
             const std::size_t n1 = Ostap::Math::round_half_down ( hh ) ;
             const std::size_t n2 = Ostap::Math::round_half_up   ( hh ) ;
@@ -359,7 +346,7 @@ namespace Ostap
             //
             return 0.5 * ( v1 + v2 ) ;
           } 
-        else if( Three == m_t )
+        else if ( Ostap::QuantileTypes::HyndmanFanType::Three == m_t )
           {
             const std::size_t nn = Ostap::Math::banker ( hh ) ;
             if ( nn ) { std::advance ( first ,  nn ) ; }	  
@@ -379,7 +366,9 @@ namespace Ostap
       // ======================================================================
     private :
       // ======================================================================
-      QuantileType m_t     { Eight } ; // quantile type
+      /// quantile type
+      Ostap::QuantileTypes::HyndmanFanType  m_t { Ostap::QuantileTypes::HyndmanFanType:: Eight } ;
+      /// check ?
       QCheck       m_check { false } ; // check 
       // ======================================================================
     };
@@ -413,6 +402,10 @@ namespace Ostap
         const double beta  = 0.4   , 
         const bool   check = false ) ;
       // ====================================================================
+      ABQuantile
+      ( const Ostap::QuantileTypes::ABQuantileType& abq          , 
+        const bool                                 check = false ) ;
+      // ====================================================================
     public:
       // ====================================================================
       template <class ITERATOR,
@@ -430,7 +423,9 @@ namespace Ostap
         if  ( 1 == N ) { return *first ; }
         if  ( 1 <= p ) { std::advance ( first , N - 1 ) ; return *first ;} 
         //
-        const double m = m_alpha + p * ( 1 - m_alpha - m_beta ) ;
+        // const double m = alpha ()  + p * ( 1 - alpha () - beta () ) ;
+        const double m = m_abq.m ( p ) ;
+        //
         const double a = p * N  + m ;
         const int    j = static_cast<int> ( std::floor ( a ) ) ;
         //
@@ -446,12 +441,17 @@ namespace Ostap
         return ( 1 - g ) * v1  + g * v2  ;
       }
       // ====================================================================
+    public:
+      // ====================================================================
+      /// get alpha 
+      inline double alpha () const { return m_abq.alpha () ; }
+      /// get beta 
+      inline double beta  () const { return m_abq.beta  () ; }
+      // ====================================================================
     private:
       // ====================================================================
-      /// parameter alpha
-      double m_alpha { 0.4   } ;
-      /// parameter beta 
-      double m_beta  { 0.4   } ;
+      /// alpha/beta keeper 
+      Ostap::QuantileTypes::ABQuantileType m_abq   {} ;
       /// check 
       QCheck m_check { false } ; 
       // ====================================================================
@@ -466,8 +466,7 @@ namespace Ostap
       // ======================================================================
     public:
       // ======================================================================
-      HarrellDavis 
-      ( const bool check = false ) ;  
+      HarrellDavis ( const bool check = false ) ;  
       // ======================================================================
     public:
       // ======================================================================
@@ -490,10 +489,11 @@ namespace Ostap
         for ( std::size_t i = 0 ; first != last ; ++first , ++i )
           {
             const double ti    = ( i + 1.0 ) / N ;
-            const double tm1   =   i * 1.0   / N ;
+            const double tp    =   i * 1.0   / N ;
             const double value = *first ;
             if ( !value ) { continue ; }
-            result += value * WHD ( N  , p , ( 1.0 + i ) / N , i * 1.0 / N ) ; 
+            //
+            result += value * WHD ( N  , p , ti , tp ) ; 
           }
         return result ;
       } ;  
@@ -510,7 +510,7 @@ namespace Ostap
        *  \f$ I_z(x,y) \f$ is normalized inncomplete beta function   
        *  @see Ostap::Math::beta_inc  
        *  - protection is added 
-       *  - cachinng is applie 
+       *  - caching is applied 
        */
       static double WHD 
       ( const std::size_t N  , 
@@ -518,13 +518,139 @@ namespace Ostap
         const double      t1 ,
         const double      t2 ) ; 
       // ======================================================================
-    } ; //                        The end of the class  Ostap::Math::HArrelavis
-    // ======================================================================== 
+    } ; //                        The end of the class  Ostap::Math::Harrelavis
+    // ========================================================================     
+    /** @class WHarrellDavis 
+     *  Harrell-Davis quatile  etimator for weighted datas 
+     *  @attention it can be  CPU espensive for large data sets
+     */
+    class WHarrellDavis : public QuantileMixin<WHarrellDavis>
+    {
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// constructor 
+      WHarrellDavis ( const bool check = false ) ;  
+      // ======================================================================
+    public:
+      // ======================================================================
+      struct Entry : public std::pair <double,double> 
+      {
+        typedef std::pair<double,double> Pair ;
+        /// constructor from the value and weight 
+        Entry
+        ( const double value  = 0 ,
+          const double weight = 1 )
+          : Pair ( value , weight ) {} 
+        /// constructor from the pair         
+        Entry ( Pair entry  ) : Pair( entry ){}
+      } ;
+      // ======================================================================
+    public : 
+      // ======================================================================
+      /** calculate the quantile using the preecalculated sum of weights and sum of squared weigths          
+       *  @param first (INPUT) start of (ordered) data-sequence 
+       *  @param last  (INPUT) end of   (ordered) data-sequence 
+       *  @param p     (input) probability 0 < p < 1 
+       *  @param sumw  (input) sum of weights (must be positive) 
+       *  @param sumw2 (input) sum of squared weights (must be positive) 
+       *  @return Harrell-Davis quantile 
+       */
+      template <class ITERATOR,
+                typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
+                typename std::enable_if<std::is_convertible<value_type,Entry>::value,bool>::type = true>
+      inline double quantile 
+      ( ITERATOR     first ,
+        ITERATOR     last  , 
+        const double p     ,
+        const double sumw  ,
+        const double sumw2 ) const
+      {
+        /// inverse total weight 
+        const double sw_inv = 1/sumw ;
+        /// the number of effectibe entries 
+        const double nstar  = sumw * sumw / sumw2 ;
+        /// effective alphas & betas
+        const double alpha  = ( nstar + 1 ) *       p   ;
+        const double beta   = ( nstar + 1 ) * ( 1 - p ) ;        
+        //
+        /// loop over all (ordered) entries 
+        double wsum   =  0 ;
+        double result =  0 ; 
+        for ( ; first != last ; ++first )
+          {
+            const Entry  entry { *first } ;
+            const double value  = entry.first   ;
+            const double weight = entry.second  ;
+            //
+            if ( !weight ) { continue ; } 
+            //
+            const double tp = std::min ( std::max ( sw_inv * wsum , 0.0 ) , 1.0 ) ;            
+            wsum += weight ;
+            const double ti = std::min ( std::max ( sw_inv * wsum , 0.0 ) , 1.0 ) ;            
+            //
+            result += value * WHD ( alpha , beta , ti , tp ) ; 
+          }
+        return result ; 
+      }
+      // ======================================================================
+      /** calculate the quantile 
+       *  @param first (INPUT) start of (ordered) data-sequence 
+       *  @param last  (INPUT) end of   (ordered) data-sequence 
+       *  @param p     (input) probability 0 < p < 1 
+       *  @return Harrell-Davis quantile 
+       */
+      template <class ITERATOR,
+                typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
+                typename std::enable_if<std::is_convertible<value_type,Entry>::value,bool>::type = true>
+      inline double quantile 
+      ( ITERATOR     first ,
+        ITERATOR     last  , 
+        const double p     ) const 
+      {
+        /// (1) calculate  sum of weights and sum of squred weights
+        double sumw  = 0 ;
+        double sumw2 = 0 ;        
+        std::for_each ( first ,
+                        last  ,
+                        [&sumw,&sumw2] ( Entry entry ) -> void
+                        {
+                          const double weight = entry.second ;
+                          if ( !weight ) { return ; } 
+                          sumw  +=          weight ;
+                          sumw2 += weight * weight ;
+                        } ) ;        
+        /// (2) calculate the quantile
+        return this->quantile ( first , last , p , sumw , sumw2 ) ;
+      }
+      // ======================================================================
+    private :
+      // ======================================================================
+      /// check ? 
+      QCheck  m_check { false } ; // check ? 
+      // ======================================================================
+    private:
+      // ======================================================================
+      /** calculate 
+       *  \f$ \ I_{t_1}(\alpha,\beta) - I_{t_2} ( \alpha, \beta) f$, where 
+       *  \f$ I_z(x,y) \f$ is normalized inncomplete beta function   
+       *  @see Ostap::Math::beta_inc  
+       *  - protection is added 
+       *  - caching is applied 
+       */
+      static double WHD 
+      ( const double alpha , 
+        const double beta  , 
+        const double t1    ,
+        const double t2    ) ; 
+      // ======================================================================
+    } ; //                      The end of the class  Ostap::Math::WHarrelavis
+    // ========================================================================             
   } //                                         The end of namespace Ostap::Math
   // ==========================================================================
 } //                                                The  end of namespace Ostap 
 // ============================================================================
-//                                                                      The END 
-// ============================================================================
 #endif // OSTAP_QUANTILES_H
+// ============================================================================
+//                                                                      The END 
 // ============================================================================