+++ source/src/Quantiles.cpp
@@ -12,6 +12,7 @@
 #include "Ostap/Hash.h"
 #include "Ostap/Math.h"
 #include "Ostap/MoreMath.h"
+#include "Ostap/QuantileTypes.h"
 #include "Ostap/Quantile.h"
 #include "Ostap/Quantiles.h"
 #include "Ostap/StatusCode.h"
@@ -26,11 +27,13 @@
 namespace 
 {
   // ===========================================================================
+  const std::size_t s_MAX_CACHE = 100000 ;
+  // ===========================================================================
   /** calculate \f$ \ I_{t_1}(\alpha,\beta) - I_{t_2} ( \alpha, \beta) f$, where 
    *  \f$ I_z(x,y) \f$ is normalized inncomplete beta function   
    *  @see Ostap::Math::beta_inc  
    *  - protection is added 
-   *  - cachinng is applie 
+   *  - caching is applied 
    */
   double _WHD_  
   ( const double alpha , 
@@ -64,13 +67,13 @@ namespace
     // 
     const auto fun2 = [alpha,beta,t1,t2]() -> double
     {  
-      const double t  = 0.5 *    ( t1 + t2 ) ;
-      const double dt = std::abs ( t2 - t1 ) ;  
-      double result  = std::log  ( dt ) ;
-      result        += ( alpha - 1 ) * std::log ( t       ) ;
-      result        += ( beta  - 1 ) * std::log ( 1 - t   ) ;
-      result        -= Ostap::Math::lnbeta ( alpha , beta ) ;
-      result         = std::exp ( result ) ;
+      const double t       = 0.5 *    ( t1 + t2 ) ;
+      const double dt      = std::abs ( t2 - t1 ) ;  
+      double       result  = std::log  ( dt ) ;
+      result              += ( alpha - 1 ) * std::log ( t       ) ;
+      result              += ( beta  - 1 ) * std::log ( 1 - t   ) ;
+      result              -= Ostap::Math::lnbeta ( alpha , beta ) ;
+      result               = std::exp ( result ) ;
       return t2 < t1 ? result : -result ;
     } ; 
     ///
@@ -79,7 +82,7 @@ namespace
     /// add result to the cache 
     {
       CACHE::Lock lock { s_cache.mutex () } ;
-      if ( 100000  < s_cache->size() ) { s_cache->clear() ; }
+      if ( s_MAX_CACHE < s_cache->size() ) { s_cache->clear() ; }
       s_cache->insert ( std::make_pair ( key , result ) ) ;
     }
     //
@@ -88,6 +91,39 @@ namespace
   // ============================================================================
 } //                                               The end of anynymous namesapce  
 // ==============================================================================
+// constructor
+// =============================================================================
+Ostap::QuantileTypes::ABQuantileType::ABQuantileType 
+( const double alpha ,
+  const double beta  )
+  : m_alpha ( alpha )
+  , m_beta  ( beta  )
+{
+  if ( s_zero  ( m_alpha     ) ) { m_alpha = 0 ; }
+  if ( s_zero  ( m_beta      ) ) { m_beta  = 0 ; }
+  if ( s_equal ( m_alpha , 1 ) ) { m_alpha = 1 ; }
+  if ( s_equal ( m_beta  , 1 ) ) { m_beta  = 1 ; }
+  //
+  Ostap::Assert ( 0 <= m_alpha && m_alpha <= 1            ,
+                  "Invalid alpha!"                        ,
+                  "Ostap::Math::ABQ"                      ,
+                  INVALID_QUANTILE , __FILE__  , __LINE__ ) ;
+  Ostap::Assert ( 0 <= m_beta && m_beta <= 1              ,
+                  "Invalid beta!"                         ,
+                  "Ostap::Math::ABQ"                      ,
+                  INVALID_QUANTILE , __FILE__  , __LINE__ ) ;
+}
+// ==============================================================================
+// constructor 
+// ==============================================================================
+Ostap::QuantileTypes::HarrellDavisType::HarrellDavisType() {} 
+// ==============================================================================
+// constructor 
+// ==============================================================================
+Ostap::QuantileTypes::P2QuantileType::P2QuantileType() {} 
+// ==============================================================================
+
+// ==============================================================================
 Ostap::Math::QCheck::QCheck
 ( const  bool check )
   :  m_check ( check)
@@ -95,27 +131,28 @@ Ostap::Math::QCheck::QCheck
 // ==============================================================================
 void Ostap::Math::QCheck::throw_exception 
 ( const char* message , 
-  const char* f , 
-  const long  l ) const 
+  const char* f       , 
+  const long  l       ) const 
 {
   Ostap::throwException 
     ( message , 
-      "Ostap::Math::QChqck" , 
+      "Ostap::Math::QCheck" , 
       INVALID_DATA , f ? f : __FILE__  , 0 <= l ? l : __LINE__ ) ;
 }
 // =============================================================================
 // constructor
 // =============================================================================
 Ostap::Math::HyndmanFan::HyndmanFan
-( const Ostap::Math::HyndmanFan::QuantileType t     , 
-  const bool                                 check ) 
+( const Ostap::QuantileTypes::HyndmanFanType  t     , 
+  const bool                                  check ) 
   : m_t      ( t    ) 
   , m_check ( check ) 
 {
-  Ostap::Assert ( One <= t  && t <= Nine , 
-		  "Invalid QuantileType!" , 
-		  "Ostap::Math::HyndmanFan" , 
-		  INVALID_QUANTILE  , __FILE__ , __LINE__ ) ;
+  Ostap::Assert ( Ostap::QuantileTypes::HyndmanFanType::One <= t  &&
+                  t <= Ostap::QuantileTypes::HyndmanFanType::Nine  , 
+                  "Invalid QuantileType!" , 
+                  "Ostap::Math::HyndmanFan" , 
+                  INVALID_QUANTILE  , __FILE__ , __LINE__ ) ;
 } 
 // ==============================================================================
 // constructor
@@ -124,23 +161,18 @@ Ostap::Math::ABQuantile::ABQuantile
 ( const double alpha , 
   const double beta  ,  
   const bool   check ) 
-  : m_alpha ( alpha )
-  , m_beta  ( beta  )
-  , m_check ( check )
-{
-  if ( s_zero  ( m_alpha    ) ) { m_alpha = 0 ; }
-  if ( s_equal ( m_beta , 1 ) ) { m_beta  = 1 ; }
-  //
-  Ostap::Assert ( 0 <= m_alpha && m_alpha <= 1            ,
-                  "Invalid alpha!"                        ,
-                  "Ostap::Math::ABQuantile"               ,
-                  INVALID_QUANTILE , __FILE__  , __LINE__ ) ;
-  Ostap::Assert ( 0 <= m_beta && m_beta <= 1              ,
-                  "Invalid beta!"                         ,
-                  "Ostap::Math::ABQuantile"               ,
-                  INVALID_QUANTILE , __FILE__  , __LINE__ ) ;
-  
-} 
+  : m_abq   ( alpha , beta )
+  , m_check ( check ) 
+{}
+// ==============================================================================
+// constructor
+// =============================================================================
+Ostap::Math::ABQuantile::ABQuantile
+( const Ostap::QuantileTypes::ABQuantileType& abq   , 
+  const bool                                  check ) 
+  : m_abq   ( abq   ) 
+  , m_check ( check ) 
+{}
 // =============================================================================
 // constructor
 // =============================================================================
@@ -150,7 +182,7 @@ Ostap::Math::HarrellDavis::HarrellDavis
 {} ;
 // =============================================================================
 /*  calculate \f$ \ I_{t_1}(\alpha,\beta) - I_{t_2} ( \alpha, \beta) f$, where 
- *  \f$ I_z(x,y) \f$ is normalized inncomplete beta function   
+ *  \f$ I_z(x,y) \f$ is normalized incomplete beta function   
  *  @see Ostap::Math::beta_inc  
  *  - protection is added 
  *  - cachinng is applie 
@@ -164,6 +196,32 @@ double Ostap::Math::HarrellDavis::WHD
 {
   const double alpha = ( N + 1 ) * p ;
   const double beta  = ( N + 1 ) * ( 1 - p ) ;
+  return _WHD_  ( alpha , beta , t1  , t2  ) ;
+}
+// ============================================================================= 
+
+
+// =============================================================================
+// constructor
+// =============================================================================
+Ostap::Math::WHarrellDavis::WHarrellDavis 
+( const bool check ) 
+  : m_check  ( check )
+{} ;
+// =============================================================================
+/*  calculate \f$ \ I_{t_1}(\alpha,\beta) - I_{t_2} ( \alpha, \beta) f$, where 
+ *  \f$ I_z(x,y) \f$ is normalized incomplete beta function   
+ *  @see Ostap::Math::beta_inc  
+ *  - protection is added 
+ *  - cachinng is applie 
+ */
+// =============================================================================
+double Ostap::Math::WHarrellDavis::WHD
+( const double alpha , 
+  const double beta  , 
+  const double t1    ,
+  const double t2    ) 
+{
   return _WHD_  ( alpha , beta , t1  , t2  ) ;
 }
 // =============================================================================